+++ mod.ts
@@ -1,4 +1,4 @@
 export { toArray } from "./collections.ts";
-export { count, cycle } from "./infinite.ts";
+export { count, cycle, repeat } from "./infinite.ts";
 export { take, takeWhile } from "./take.ts";
 export { assertStreams, assertStreamStartsWith } from "./testing.ts";