+++ CHANGES.md
@@ -10,6 +10,7 @@ To be released.
 
  -  Added *mod.ts* module which re-exports everything from other modules.
  -  Added `cycle()` function.
+ -  Added `repeat()` function.
 
 
 Version 0.1.0